flask-app/requirements.txt
aniso8601==9.0.1
astroid==2.5.3
attrs==21.4.0
black==22.1.0
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
Flask==2.0.3
flask-expects-json==1.7.0
Flask-JWT==0.3.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.0
PyJWT==1.4.2
pylint==2.7.4
pyrsistent==0.18.1
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.12.1
